+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - add Blargg `halt_bug` DMG screenshot and update Blargg test documentation
+- include each test suite's readme file(s)
 
 
 ## [v3.1](https://github.com/c-sp/gameboy-test-roms/releases/tag/v3.1) *(2021-09-30)*